Users can authenticate with Twitter
This commit handles both sign in and sign up.

The `authenticate` function is a magic callback that gets called after the user
authenticates through Twitter. It receives all of the information from Twitter
in a Creds object, which has a relevant information (screen name, oauth token,
oauth secret, etc) in the `credsExtra` field.

When `authenticate` receives the Twitter credentials, it sets `sessionKey`
(which is stored on App) in the session to the Twitter ID. It doesn't use the
Twitter name because that might change. Then it runs `authenticateUser`, which
finds or creates a user based on the credentials.

When we need to find a user (with `requireAuth`, since the authorization rules
guarantee a user is signed in for the MonikerR route), it uses `maybeAuthId`,
which searches the session for `sessionKey` and uses it to find a user.