Defining our own maybeAuthId isn't necessary
The [minimal complete definition][0] for YesodAuth is:

`loginDest, logoutDest, (authenticate | getAuthId), authPlugins, authHttpManager`

Since that list _doesn't_ include `maybeAuthId`, we can assume that implementing
`authenticate` (and the rest of the functions) defines a sane `maybeAuthId` for
us. And in fact it does!  Everything Just Works! Wow!

We also don't need to do anything with the session since we're not passing any
data between `authenticate` and `maybeAuthId`, so `sessionKey` in Application
can go away as well as a bunch of session-related code.

[0]: https://hackage.haskell.org/package/yesod-auth-1.4.12/docs/Yesod-Auth.html#t:YesodAuth