Update moniker if it's today *in their timezone*
The moniker's day is stored without a time zone, as a pure date like 2016-2-21.

We interpret it as being in the owning user's timezone.

Here is the process for determining if a moniker should be sent to Twitter:

* Take the current UTC time and, for each user, convert it to that user's
  timezone.
* Using that time, find out what "today" is for that user.
* If "today" for the user is the same as the moniker's day, then the moniker
  should be updated.

One wrinkle is that, since Postgres doesn't know anything about timezones, we
find monikers for UTC yesterday/UTC today/UTC tomorrow since "UTC tomorrow" may
be "the user's today". We get some extra monikers that won't be updated this
way, but it's easier to filter in Haskell than in Haskell _and_ Postgres.