Display the date dropdown in the user's timezone
We were displaying the dropdown in UTC, so if it was the 22nd for the user and
the 23rd in UTC, the dropdown would default to the 23rd. The backend logic that
ensured future dates also used UTC, so if it was already "tomorrow" in UTC,
users couldn't enter a name.

Now the code is aware of the user's timezone and uses it everywhere.

Refactoring:

* The `Handler.TodaysMonikersTask.userDay` method was folded into
  `Croniker.Time.localToday`.