This function can find the Yesod App by itself
Since `updateName` runs in `Handler`, it can call `getYesod` and we don't need
to pass it in from `updateTodaysMonikers`.

Also, instead of grabbing every field in `App`, we only grab what we need: the
Twitter consumer key and secret.

The `App{twitterConsumerKey, twitterConsumerSecret}` syntax requires the
`NamedFieldPuns` language extension. Without the extension, it looks like this:

    App{twitterConsumerKey=twitterConsumerKey, twitterConsumerSecret=twitterConsumerSecret}