Add specs for normalizing monikers
`normalize` is the function that turns a user-inputted moniker into the version
that Twitter verifies. Think of it as a filter that runs before any of the
checking runs.

For example, `normalize` strips whitespace from the beginning and end. So a
function that checks length can check the length without having to do that
itself, and " hello  " will be correctly read as a 5-character name.