Check that a name doesn't contain whitespace
`normalize` strips whitespace from the beginning and end of the moniker.

But consider a moniker like `one \t\n one`. It has special whitespace in the
middle, so it's not stripped, and Twitter will reject the moniker.

Now Croniker checks for that and rejects the name, just like Twitter would.