Use intermediate data structure for Profile form
A `Profile` can have an (optional) Base64-encoded picture. The user uploads a
picture, Croniker Base64-encodes its bytes, and stores those encoded bytes in
the database.

However, an uploaded picture is represented as a `FileInfo`. `Profile` wants a
`Maybe Text` for its `picture` field, and so its applicative `Form` must turn
the `FileInfo` into a `Maybe Text` or else it won't compile.

Turning that `FileInfo` into the actual uploaded bytes is unsafe, since the
uploaded file may have been deleted on the server before Yesod can process it.
Since the operation is unsafe, it's very difficult (impossible?) to pull out the
bytes in an applicative `Form`.

The solution: don't build the form around `Profile` -- build it around an
easier-to-use type instead, then post-process into a `Profile`. That's exactly
what `FormProfile` is. It accepts a `FileInfo`, so Croniker doesn't need to do
complex unsafe operations in the applicative `Form`. Once the form is
successfully uploaded, `addProfile` turns the `FormProfile` into a `Profile` and
inserts it into the database. Since `addProfile` runs in a `Handler`, it can
perform unsafe operations like Base64-encoding the uploaded file, then hand the
Base64-encoded bytes to `Profile` before it's inserted into the database.

Thank you to:

* https://github.com/rab206/fp-cv-file-server/blob/d9a7fb5446cde68ade71b31abc35a4d5c1abfe34/Foundation.hs
* https://github.com/rab206/fp-cv-file-server/blob/d9a7fb5446cde68ade71b31abc35a4d5c1abfe34/Handler/Home.hs