Use applicative-style form for Profile
The monadic style wasn't necessary and we can `fmap` the normalization functions
through the `areq`/`aopt` functions.

Of note is the `Nothing <$ aopt fileField ...`, which always sets the result of
the file field to `Nothing`. `withPossibleProfilePicture` will set the
file field to an actual Base64-encoded value later in the POST handler.