Show a friendly error if a date is already taken
Previously, attempting to schedule a Profile on a date that already had a
Profile would show a 500 error with a SQL error (since there's a unique index in
the DB).

Now it shows an error message instead, which lets the user know what's happening
and how to fix it.

Fixes https://github.com/gabebw/croniker/issues/53