Ensure that URLs are really URLs
The previous solution's simple parsing meant that names like `adarsh.tar.gz`
(which is not a valid URL) was interpreted as a URL.

Now we use the tld library to ensure that only real URLs are parsed as such.

We must add `http://` to each possible URL before parsing it or else
`parseTLDText` always returns `Nothing`.

Fixes #52.