Default to the next available date
The "Date" field now defaults to the first day in the next 365 days (not
including today) that does _not_ have a profile scheduled. For example:

* if nothing is scheduled for today, it defaults to tomorrow.
* if nothing is scheduled for tomorrow, it defaults to tomorrow.
* if there are profiles scheduled for the next 7 days, it defaults to 8 days from now.

If the next 365 days are all taken, it falls back to tomorrow and the user will
have to manually select a date from the date picker.

The SQL uses `generate_series` to generate a temporary table containing the next
365 days, does a `LEFT OUTER JOIN` with `profile.date`, then selects the first
generated date where `profile.date` is NULL (meaning that there's no profile
scheduled for that day).