Display profiles based on the user's local date
Previously, the profiles were displayed based on `today`, which was in UTC. This
meant that if a user in Pacific time viewed the homepage towards the end of the
day on June 16, then "today" was actually June 17, and the profile for June 16
was not displayed.

Now, by using `localToday`, today's profile is always displayed.

Additionally, the method has been renamed to better communicate what it does.