Enforce a more explicit grid on the profiles index
The picture and the moniker are next to each other, but only the picture was in
a grid column. This means that as more fields are added, they could flow around
the picture since they won't be in a grid column.

Now the picture is in its own column and the text fields are in their own
column, so the text won't ever flow around the profile picture.

When there is no profile picture, the text (right now just the moniker) takes up
all of the space available.

If there is a profile picture, the picture takes 1 column and the text takes the
rest.