Add created_at/updated_at fields to all tables
The `MigrationOnly` notation in `config/models` means that Haskell code will
entirely ignore the field (except to create it when the app starts up). It's
present in the database, but not used at all in Yesod. This is perfect for
`created_at` and `updated_at`, which are set automatically via (respectively) a
default value and a database trigger.

`MigrationOnly` docs: https://github.com/yesodweb/persistent/wiki/Persistent-entity-syntax#migrationonly

Fixes #77.