Rework how validations are shown to the user
Previously, monikers were normalized, then validated. This meant that if a user
entered an entirely invalid moniker like "<<<<<", it would get normalized to ""
(an empty string), and the user would see an error about not allowing blank
monikers...but they'd still see "<<<<<" in the form input.

This was confusing: the user _did_ enter something, but it looked like Croniker
ate it and is showing an incorrect warning. In other words, it looked like
Croniker's fault, not the user's.

To fix this, don't ever normalize the moniker before validating. Instead, show
targeted and helpful error messages like "Twitter doesn't allow < or >" that
will help users craft clever and valid monikers.

There is a little loss of functionality with this change: monikers are not
stripped of whitespace before being validated or saved. Thus, if a user enters
`  twitter  ` (which is invalid because it contains "twitter"), the user will
see the input with errors and with `  twitter  ` still in the input. They won't
know that they could gain back a few characters by removing that whitespace. I
think that's a reasonable tradeoff to make, since:

1) it's hard to figure out how to change the value inside the input;
2) the new error messages are much clearer