Rework how validations are shown to the user
Previously, monikers were normalized, then validated. This meant that if a user
entered an entirely invalid moniker like "<<<<<", it would get normalized to ""
(an empty string), and the user would see an error about not allowing blank
monikers...but they'd still see "<<<<<" in the form input.

This was confusing: the user _did_ enter something, but it looks like Croniker
ate it and is showing an incorrect warning. It looks like Croniker's fault, not
the user's.

To fix this, don't ever normalize the moniker before validating. Instead, show
targeted and helpful error messages like "Twitter doesn't allow < or >" that
will help users craft clever and valid monikers.