Use a Dockerfile that can actually run the app
Alpine Linux is based on musl, while GHC-built binaries rely on glibc.
Therefore there's a bunch of weird errors when trying to run the image,
like "__sprintf_chk: symbol not found".

Therefore, use a glibc-based distribution like Ubuntu.

The resulting image is 332MB (Ubuntu) instead of 150MB (Alpine), but it
can actually run the programs, which is a huge point in Ubuntu's favor.

Here's a glibc-related issue on Alpine Linux:
https://github.com/gliderlabs/docker-alpine/issues/149