Redirect correctly from unauthorized paths
Because of the change to `loginHandler` in
eb303cb080b0652c8df7926d69d4920550db571e, visiting a page the requires
authentication (like `/profiles`) when signed out redirects to
`/auth/login`, which is an ugly unstyled page.

Instead, we want to redirect to `/`, which will show a pretty sign-in
button to signed-out users.