Cache the result of Docker builds
Please note: the "real" way to do this is to use Docker Layer Caching
(DLC). DLC is built in to developer-facing tools like Docker for Mac,
but Circle CI requires a paid account and charges extra for it. Since
this is a personal project, it's not worth it to pay for DLC, so this is
the free solution. Documentation is [here][dlc].

OK, now let's explain what's going on.

For the two targets (the two `FROM` statements) in the Dockerfile, the
following happens:

* Restore a cache of the target, if it exists (with `docker load`)
* Build the target, using the restored cache (using `--cache-from`)
* Save a cache of the target (with `docker save`)

But why build `croniker:fpco` and `croniker:final` separately?  Because
if I build the whole Dockerfile as one piece, `docker save` will save
only the final result, which is the `croniker:final` image.
`croniker:fpco` won't be saved at all. Therefore, I need to target
`fpco`, and save the result of that intermediate build, then build the
rest of it. (This issue happens because it's a [multi-stage
build][multistage], which I use because it makes the final image that
gets uploaded to Heroku much much smaller.)

By using this load/save pattern, I save 12 minutes per deploy, which is
about 25%. (And that savings happens even though restoring the 4GB cache
takes 10 minutes.)

* No caching (42 minutes): https://circleci.com/gh/gabebw/croniker/293
* With caching (30 minutes): https://circleci.com/gh/gabebw/croniker/313

These articles were very helpful for understanding how to cache
multi-stage builds:

* https://andrewlock.net/caching-docker-layers-on-serverless-build-hosts-with-multi-stage-builds---target,-and---cache-from/
* https://circleci.com/blog/how-to-build-a-docker-image-on-circleci-2-0/

[dlc]: https://circleci.com/docs/2.0/docker-layer-caching/
[multistage]: https://docs.docker.com/develop/develop-images/multistage-build/